separating firebase helper functions into files
