--- lib/firebase-helpers/private/general.ts
@@ -1,5 +1,0 @@
-export const verifyServerSide = () => {
-  if (typeof window !== "undefined") {
-    throw new Error("This function can only be called on the server");
-  }
-};
