+++ pages/admin/directory.tsx
@@ -18,18 +18,14 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog";
 
-import {
-  deleteDocument,
-  deleteReferences,
-} from "@/lib/firebase-helpers/public/directory";
+import { deleteDocument } from "@/lib/firebase-helpers/general";
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
   DocumentData,
   MemberEmail,
   MemberPublic,
   RegionPublic,
-  getFirebaseTable,
-} from "@/lib/firebase-helpers/api";
+} from "@/lib/firebase-helpers/interfaces";
 import {
   CompanySizeEnum,
   FirebaseTablesEnum,
@@ -49,7 +45,8 @@ import { FC, ReactNode, useEffect, useState } from "react";
 import { useAuthState } from "react-firebase-hooks/auth";
 import { signInWithGoogle, signOutWithGoogle } from "../../lib/firebase";
 import { MemberEdit } from "@/components/MemberEdit";
-import { getAllMemberReferencesToDelete } from "@/lib/firebase-helpers/public/directory";
+import { getAllMemberReferencesToDelete } from "@/lib/firebase-helpers/members";
+import { deleteReferences } from "@/lib/firebase-helpers/members";
 
 export async function getStaticProps() {
   return {
