+++ pages/join/thank-you.tsx
@@ -20,8 +20,8 @@ import { cn } from "@/lib/utils";
 import theme from "@/styles/theme";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { buttonVariants } from "@/components/ui/button";
-import { Filter, getFilters } from "@/lib/firebase-helpers/api";
-import { getMembers } from "@/lib/firebase-helpers/private/members";
+import { Filter, getFilters } from "@/lib/firebase-helpers/interfaces";
+import { getMembers } from "@/lib/firebase-helpers/members";
 
 export async function getStaticProps() {
   const { members, focuses, industries } = await getMembers();
